Trigger a binder env build when deploying doc
The goal is to shorten the waiting time for the user following a
binder link to a tutorial.

By doing this triggering, the build process is already launched on each
BinderHub deployments. Thus, when requesting this binder environment
later, it is known as already built and the image has just to be pulled
instead of being created.