Create new DecomonLayer base class to be used for bounds propagation
The same layer class is now used for forward and backward propagation.

The goal is to simplified the implementation of the decomon version
of a custom layer for the user wanting to use decomon algorithms on a
model involving such a custom layer.

When implementing a new decomon layer it is sufficient to implement
- get_affine_bounds(): returns affine relaxation of the keras layer
- get_ibp_bounds(): returns constant relaxation of the keras layer

In the case of a linear keras layer, one can set the class attribute
`linear` to True and then implement get_affine_representation() instead
of get_affine_bounds(). Some computations will be simplified in this
case. In particular, no oracle bounds are needed to propagate affine
bounds.

One can also, for performance reasons, override directly
forward_affine_propagate() and backward_affine_propagate() instead of
implementing get_affine_bounds().

The main attributes of the decomon layers are
- layer: the keras underlying layer
- perturbation_domain: on which we propagate bounds
- ibp: do we propagate constant bounds (forward only)
- affine: do we propagate affine bounds (meaningless for backward)
- propagation: forward or backward, the direction of bounds propagation

ibp and affine booleans replace the previous modes as they seem to be
the relevant variables when testing inside decomon layer methods;