Change (flatten) inputs/outputs format for DecomonLayer.call()
To be consistent with DecomonModel that should be seen as special
DecomonLayer and that, as functional keras model, can not be defined
with inputs and outputs being nested lists of tensors, we choose to
flatten inputs and outputs for DecomonLayer.call().

In other methods, that will be used by custom decomon layers developers
we keep the split into affine, constant and model inputs.

We manage all this split/flatten stuff necessary to go from call()
format to call_backward/call_forward format, by a dedicated class
InputsOutpusSpec.

We update inputs generation utility functions used in tests accordingly.