Add plot_model() utility
We tweak keras.utils.plot_model() to be able to represent some
decomon_layer attributes.
By default we add the propagation and the underlying keras layer name
We can add custom attributes thanks to the argument
show_layer_attributes.

An example:

        show_layer_attributes = {
            "propagation": {
                "forward": {
                    "bgcolor": "#b3e6ff",
                },
                "backward": {
                    "bgcolor": "#ffccdd",
                },
                None: {"color": "black"},
            },
            "layer.name": {},
        }

We list the attribute (even nested ones) to show,
and the style (bgcolor and color) to apply for them.

We define a default style by adding a None entry into the style dict

		None: "color": "black" -> all propagation will be written in black

We define a style for a specific value by adding an entry with the value
as a key and the style dict as value.

                "forward": {
                    "bgcolor": "#b3e6ff",
                },