Add LinkToPerturbationDomainInput layer to use in clone() for fully linear keras models
When a keras model is fully linear, its affine bounds will be
independent from sample, and in backward prapagation with no
first backward_bounds, they will even be eagerly computed as backend
tensors.
It is not possible to instantiate a decomon model in that case since
some outputs are not depending on the perturbation domain input.

So we artificially make them depend on the perturbation domain input
by adding a dedicated layer at the end of the conversion.