Improve backward_bounds management
- Create a dedicated layer to preprocess backward_bounds
- Change preferred format for backward_bounds to a flatten list
which is the concatenation of backward_bounds for each model output.
The reason being that this is how a crown model will return backward
bounds and thus such crown models can be directly used to generate
backward bounds.
- If None, or empty or list tof empty list we set backward_bounds to None
to be used by convert.
- Else we preprocess them with BackwardInput to ensure having a list of 4
* nb_model_outputs. (See conventions taken in BackwardInput docstring)
- convert() get a flatten list to be used by Fuse layer, but we need
to split it in convert_backward to feed crown_model  with the expected
list of list of tensors.
- The flatten list of backward_bounds *before* preprocessing is used as
inputs for DecomonModel (together with perturbation domain input).
- We manage possibility to have backward_bounds without batchsize (if
  coming from a linear model) with from_linear_backward_bounds parameter.
  We use it for:
  - Fuse layer
  - BackwardInput layer
  - convert_backward()/crown_model() in order to compute correctly the
    model output shape to consider (derived from the backward bounds if
    not empty)